Legger til populering av barna i requesten.
